% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basie.R
\name{basie}
\alias{basie}
\title{Computes the BASIE (BAyeSian Interpretation of Estimates) posterior
distribution}
\value{
A list containing the mean and standard deviation of the BASIE
posterior distribution.
}
\description{
Computes the BASIE (BAyeSian Interpretation of Estimates) posterior
distribution

Computes the BASIE (BAyeSian Interpretation of Estimates) posterior
distribution
}
\details{
Implementation of \href{https://www.mathematica.org/download-media?MediaItemId=\%7B68A84423-AED0-4349-8524-51E91336507C\%7D}{BASIE}
}
\examples{
priorMean <- 0
priorSD <- 0.03
likMean <- 0.071
likSD <- 0.074

basie_estimate <- basie$new(
  priorMean = priorMean,
  priorSD = priorSD,
  likMean = likMean,
  likSD = likSD
)

basie_estimate$vizdraws(
  percentage = TRUE,
  breaks = c(0, 0.05),
  break_names = c("Worse", "Equivalent", "Better"),
  height = 450,
  width = 750,
  display_mode_name = TRUE
)

txt <-
  glue::glue(
    "This example assumes that the prior distribution mean is ",
    "{scales::percent(priorMean)}, and its standard deviation ",
    "is {scales::percent(priorSD)}. ",
    "Furthermore, we imagine a study that found ",
    "a point estimate for the effect of {scales::percent(likMean)} ",
    "with a standard error of {scales::percent(likSD)}. ",
    "Finally, we assume that different decisions would be made ",
    "if lift is negative, positive but less than 5\%, or greater than 5\%."
  )

cat(stringr::str_wrap(txt), "\n")

basie_estimate$plotProbabilities()

}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mean}}{Mean of the BASIE posterior distribution.}

\item{\code{sd}}{Standard deviation of the BASIE posterior distribution.}

\item{\code{prior}}{Prior distribution.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-basie-new}{\code{basie$new()}}
\item \href{#method-basie-vizdraws}{\code{basie$vizdraws()}}
\item \href{#method-basie-probability}{\code{basie$probability()}}
\item \href{#method-basie-plotProbabilities}{\code{basie$plotProbabilities()}}
\item \href{#method-basie-clone}{\code{basie$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basie-new"></a>}}
\if{latex}{\out{\hypertarget{method-basie-new}{}}}
\subsection{Method \code{new()}}{
Estimate the BASIE posterior distribution.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basie$new(priorMean, priorSD, likMean, likSD)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{priorMean}}{Prior distribution mean.}

\item{\code{priorSD}}{Prior distribution standard deviation.}

\item{\code{likMean}}{Likelihood mean (point estimate).}

\item{\code{likSD}}{Likelihood standard deviation (standard error of the point
estimate).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{Basie}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basie-vizdraws"></a>}}
\if{latex}{\out{\hypertarget{method-basie-vizdraws}{}}}
\subsection{Method \code{vizdraws()}}{
Plot the prior and BASIE posterior distributions.
See \code{\link[vizdraws:vizdraws]{vizdraws::vizdraws()}} for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basie$vizdraws(draws = 50000L, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{draws}}{Number of draws for the posterior.}

\item{\code{...}}{Other arguments passed to \code{vizdraws::vizdraws()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basie-probability"></a>}}
\if{latex}{\out{\hypertarget{method-basie-probability}{}}}
\subsection{Method \code{probability()}}{
Estimates the probability that the effect of the intervention is
greater than \code{x}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basie$probability(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Threshold of interest.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basie-plotProbabilities"></a>}}
\if{latex}{\out{\hypertarget{method-basie-plotProbabilities}{}}}
\subsection{Method \code{plotProbabilities()}}{
Plot the probability that the effect of the intervention is greater
than a given threshold.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basie$plotProbabilities(from = -0.08, to = 0.08, length.out = 100, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{from}}{The starting value for the x-axis.}

\item{\code{to}}{The ending value for the x-axis.}

\item{\code{length.out}}{The number of points to generate for the x-axis.}

\item{\code{...}}{Other arguments passed to \code{ggplot2::ggplot()}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basie-clone"></a>}}
\if{latex}{\out{\hypertarget{method-basie-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basie$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
