% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/did.R
\name{calculateDIDEffect}
\alias{calculateDIDEffect}
\title{Calculate Difference-in-Differences Effect}
\usage{
calculateDIDEffect(
  mean_pre_control,
  sd_pre_control,
  n_pre_control,
  mean_post_control,
  sd_post_control,
  n_post_control,
  mean_pre_treat,
  sd_pre_treat,
  n_pre_treat,
  mean_post_treat,
  sd_post_treat,
  n_post_treat
)
}
\arguments{
\item{mean_pre_control}{Mean outcome for the control group in the pre-treatment period.}

\item{sd_pre_control}{Standard deviation of the outcome for the control group in the pre-treatment period.}

\item{n_pre_control}{Sample size for the control group in the pre-treatment period.}

\item{mean_post_control}{Mean outcome for the control group in the post-treatment period.}

\item{sd_post_control}{Standard deviation of the outcome for the control group in the post-treatment period.}

\item{n_post_control}{Sample size for the control group in the post-treatment period.}

\item{mean_pre_treat}{Mean outcome for the treatment group in the pre-treatment period.}

\item{sd_pre_treat}{Standard deviation of the outcome for the treatment group in the pre-treatment period.}

\item{n_pre_treat}{Sample size for the treatment group in the pre-treatment period.}

\item{mean_post_treat}{Mean outcome for the treatment group in the post-treatment period.}

\item{sd_post_treat}{Standard deviation of the outcome for the treatment group in the post-treatment period.}

\item{n_post_treat}{Sample size for the treatment group in the post-treatment period.}
}
\value{
A list containing the following components:
\item{did_estimate}{The DID estimate of the average treatment effect.}
\item{se_did}{The standard error of the DID estimate.}
}
\description{
This function calculates the difference-in-differences (DID) estimate of the
average treatment effect, along with its standard error. It assumes a
simple DID design with two groups (control and treatment) and two time
periods (pre and post).
}
\examples{
\donttest{
calculateDIDEffect(
  mean_pre_control = 10, sd_pre_control = 2, n_pre_control = 50,
  mean_post_control = 12, sd_post_control = 2.5, n_post_control = 50,
  mean_pre_treat = 11, sd_pre_treat = 2.1, n_pre_treat = 50,
  mean_post_treat = 15, sd_post_treat = 2.6, n_post_treat = 50
)
}

}
