% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBaseline.R
\name{checkBaseline}
\alias{checkBaseline}
\title{Check Baseline Equivalency.}
\usage{
checkBaseline(data, variables, treatment)
}
\arguments{
\item{data}{dataframe with the pre-intervention variables,
and the treatment indicator.}

\item{variables}{vector of with the names of the pre-intervention variables.}

\item{treatment}{name of the treatment indicator.}
}
\value{
tibble with the standardized difference of the pre-intervention
variables. The tibble includes
variables: the variable name,
std_diff: the standardized difference for that variable as a number
balance: the standardized difference for that variable as a factor
variable.
For more details about this methodology check
\url{https://ies.ed.gov/ncee/wwc/Docs/OnlineTraining/wwc_training_m3.pdf}.
}
\description{
Check Baseline Equivalency.
}
