% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMissing.R
\name{countMissing}
\alias{countMissing}
\title{Count missing values (NA) in a dataframe}
\usage{
countMissing(df)
}
\arguments{
\item{df}{A dataframe to analyze.}
}
\value{
A tibble with the following columns:
\itemize{
\item NA_<column_name>: Number of NAs in each original column
\item missing_rows: Number of rows with at least one NA across all columns
}
}
\description{
This function takes a dataframe as input and returns a tibble
summarizing the number of missing values (NA) in each column and
the number of rows with at least one missing value.
}
