% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizer.R
\name{.combineColumns}
\alias{.combineColumns}
\title{Combine and Unite Columns}
\usage{
.combineColumns(df)
}
\arguments{
\item{df}{A data frame containing the columns to be combined.}
}
\value{
A data frame with the combined 'group' column and specified columns.
}
\description{
This function takes a data frame, identifies columns that are not specified
in the exclusion list, and combines them into a new column called 'group'.
The original columns used for the combination are then removed.
Finally, it returns a data frame with only the 'group',
'variables', 'std_diff', and 'balance' columns.
}
