% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStanParameter.R
\name{getStanParameter}
\alias{getStanParameter}
\title{Extracts parameter from Stan model.}
\usage{
getStanParameter(fit, par)
}
\arguments{
\item{fit}{A Stan fit model.}

\item{par}{Name of the parameter you want to extract.}
}
\value{
A long tibble with draws for the pamete.
}
\description{
Extracts parameter from Stan model.
}
