% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdleLogNormal.R
\docType{class}
\name{hurdleLogNormal}
\alias{hurdleLogNormal}
\title{Bayesian Hurdle Log-Normal Model Factory}
\description{
A class for creating and managing Bayesian Hurdle Log-Normal Models
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{version}}{imt package version used to fit model}

\item{\code{ATE_draws}}{Posterior draws for the Average Treatment Effect on the positive outcome}

\item{\code{tau_prob_zero_draws}}{Posterior draws for the change in probability of zero outcome due to treatment}

\item{\code{mcmChecks}}{MCMC diagnostics}

\item{\code{credible_interval}}{Credible interval for the treatment effect}

\item{\code{prior_ATE}}{Prior distribution for ATE}

\item{\code{prior_tau_prob_zero}}{Prior distribution for tau_prob_zero}

\item{\code{predict_list}}{List of predictions}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-hurdleLogNormal-new}{\code{hurdleLogNormal$new()}}
\item \href{#method-hurdleLogNormal-tracePlot}{\code{hurdleLogNormal$tracePlot()}}
\item \href{#method-hurdleLogNormal-calcProb}{\code{hurdleLogNormal$calcProb()}}
\item \href{#method-hurdleLogNormal-pointEstimate}{\code{hurdleLogNormal$pointEstimate()}}
\item \href{#method-hurdleLogNormal-credibleInterval}{\code{hurdleLogNormal$credibleInterval()}}
\item \href{#method-hurdleLogNormal-plotPrior}{\code{hurdleLogNormal$plotPrior()}}
\item \href{#method-hurdleLogNormal-posteriorPredictiveCheck}{\code{hurdleLogNormal$posteriorPredictiveCheck()}}
\item \href{#method-hurdleLogNormal-clone}{\code{hurdleLogNormal$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-new"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-new}{}}}
\subsection{Method \code{new()}}{
Get the package version


Get the posterior draws for ATE


Get the posterior draws for change in zero probability


Get the MCMC diagnostics


Get the credible interval


Get the prior for ATE


Get the prior for change in zero probability


Get the list of predictions


Create a new Bayesian Hurdle Log-Normal Model object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$new(
  data,
  y,
  x,
  treatment,
  mean_alpha_logit = -3,
  sd_alpha_logit = 2,
  mean_beta_logit = NULL,
  sd_beta_logit = NULL,
  tau_mean_logit = 0,
  tau_sd_logit = 0.5,
  mean_tau = 0,
  sigma_tau = 0.035,
  seed = 1982,
  fit = TRUE,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data frame to be used}

\item{\code{y}}{Name of the outcome variable in the data frame}

\item{\code{x}}{Vector of names of all covariates in the data frame}

\item{\code{treatment}}{Name of the treatment indicator variable in the data frame}

\item{\code{mean_alpha_logit}}{Prior mean for alpha in the logit part (default: -3)}

\item{\code{sd_alpha_logit}}{Prior standard deviation for alpha in the logit part (default: 2)}

\item{\code{mean_beta_logit}}{Prior mean for beta in the logit part (default: 0 vector)}

\item{\code{sd_beta_logit}}{Prior standard deviation for beta in the logit part (default: 0.5 vector)}

\item{\code{tau_mean_logit}}{Prior mean for the treatment effect in the logit part (default: 0)}

\item{\code{tau_sd_logit}}{Prior standard deviation for the treatment effect in the logit part (default: 0.5)}

\item{\code{mean_tau}}{Prior mean for the treatment effect in the log-normal part (default: 0)}

\item{\code{sigma_tau}}{Prior standard deviation for the treatment effect in the log-normal part (default: 0.035)}

\item{\code{seed}}{Seed for Stan fitting}

\item{\code{fit}}{Flag for fitting the data to the model or not}

\item{\code{...}}{Additional arguments for Stan}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisible
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-tracePlot"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-tracePlot}{}}}
\subsection{Method \code{tracePlot()}}{
Plot MCMC trace for the ATE and tau_prob_zero parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$tracePlot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional arguments for Stan}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-calcProb"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-calcProb}{}}}
\subsection{Method \code{calcProb()}}{
Calculates the posterior probability of an effect being greater than, less than,
or within a range defined by thresholds.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$calcProb(effect_type, a = 0, b = NULL, prior = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect_type}}{The type of effect to calculate probability for ("ATE" or "tau_prob_zero")}

\item{\code{a}}{Optional. Lower bound for the threshold.}

\item{\code{b}}{Optional. Upper bound for the threshold.}

\item{\code{prior}}{Logical. If TRUE, calculates probabilities based on
the prior distribution. If FALSE (default), uses the posterior distribution.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string summarizing the estimated probability
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-pointEstimate"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-pointEstimate}{}}}
\subsection{Method \code{pointEstimate()}}{
Calculate point estimate of the effect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$pointEstimate(effect_type, median = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect_type}}{The type of effect to calculate the point estimate for ("ATE" or "tau_prob_zero")}

\item{\code{median}}{Logical value. If TRUE (default), the median of
the draws is returned. If FALSE, the mean is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value representing the point estimate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-credibleInterval"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-credibleInterval}{}}}
\subsection{Method \code{credibleInterval()}}{
Calculates credible interval for the effect of the intervention
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$credibleInterval(effect_type, width = 0.95, round = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{effect_type}}{The type of effect to calculate the credible interval for ("ATE" or "tau_prob_zero")}

\item{\code{width}}{Numeric value between 0 and 1 representing the desired
width of the credible interval (e.g., 0.95 for a 95\% credible interval).}

\item{\code{round}}{Integer value indicating the number of decimal places to round
the lower and upper bounds of the credible interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string with the following information:
\itemize{
\item The probability associated with the specified width
\item The lower and upper bounds of the credible interval, rounded to the
specified number of decimal places
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-plotPrior"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-plotPrior}{}}}
\subsection{Method \code{plotPrior()}}{
Plot prior distributions for ATE and tau_prob_zero
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$plotPrior(bins = 2000, xlim_ate = NULL, xlim_tau = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bins}}{Number of bins for the histograms}

\item{\code{xlim_ate}}{Optional. Limits for the x-axis of the ATE histogram}

\item{\code{xlim_tau}}{Optional. Limits for the x-axis of the tau_prob_zero histogram}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list containing two ggplot objects: one for ATE and one for tau_prob_zero
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-posteriorPredictiveCheck"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-posteriorPredictiveCheck}{}}}
\subsection{Method \code{posteriorPredictiveCheck()}}{
Performs Posterior Predictive Checks and generates a density overlay plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$posteriorPredictiveCheck(n = 50, xlim = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{Number of posterior predictive samples to use in the plot}

\item{\code{xlim}}{Optional. Limits for the x-axis of the plot}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A ggplot object representing the density overlay plot
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hurdleLogNormal-clone"></a>}}
\if{latex}{\out{\hypertarget{method-hurdleLogNormal-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hurdleLogNormal$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
