% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/im-package.R
\docType{package}
\name{im-package}
\alias{im}
\alias{im-package}
\title{The 'im' package}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\author{
\strong{Maintainer}: Ignacio Martinez \email{martinezig@google.com} (\href{https://orcid.org/0000-0002-3721-8172}{ORCID})

}
\keyword{internal}
