% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imt-package.R
\docType{package}
\name{imt-package}
\alias{imt}
\alias{imt-package}
\title{The 'imt' package}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (NA). RStan: the R interface to Stan. R package version 2.32.6. https://mc-stan.org
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/google/imt}
  \item Report bugs at \url{https://github.com/google/imt/issues}
}

}
\author{
\strong{Maintainer}: Ignacio Martinez \email{martinezig@google.com} (\href{https://orcid.org/0000-0002-3721-8172}{ORCID})

}
\keyword{internal}
